@@ -1,15 +1,8 @@
 import os
-import re
-import sys
-from setuptools import setup, find_packages
 
+from setuptools import find_packages, setup
 
-install_requires = ['pyodbc']
-
-PY_VER = sys.version_info
-
-if not PY_VER >= (3, 5):
-    raise RuntimeError("aioodbc doesn't support Python earlier than 3.5")
+install_requires = ["pyodbc>=5.0.1"]
 
 
 def read(f):
@@ -19,47 +12,54 @@ def read(f):
 extras_require = {}
 
 
-def read_version():
-    regexp = re.compile(r"^__version__\W*=\W*'([\d.abrc]+)'")
-    init_py = os.path.join(os.path.dirname(__file__),
-                           'aioodbc', '__init__.py')
-    with open(init_py) as f:
-        for line in f:
-            match = regexp.match(line)
-            if match is not None:
-                return match.group(1)
-        else:
-            raise RuntimeError('Cannot find version in aioodbc/__init__.py')
-
-
 classifiers = [
-    'License :: OSI Approved :: Apache Software License',
-    'Intended Audience :: Developers',
-    'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.5',
-    'Programming Language :: Python :: 3.6',
-    'Operating System :: POSIX',
-    'Environment :: Web Environment',
-    'Development Status :: 3 - Alpha',
-    'Topic :: Database',
-    'Topic :: Database :: Front-Ends',
-    'Framework :: AsyncIO',
+    "License :: OSI Approved :: Apache Software License",
+    "Intended Audience :: Developers",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Operating System :: POSIX",
+    "Environment :: Web Environment",
+    "Development Status :: 3 - Alpha",
+    "Topic :: Database",
+    "Topic :: Database :: Front-Ends",
+    "Framework :: AsyncIO",
 ]
 
+project_urls = {
+    "Website": "https://github.com/jettify/uddsketch",
+    "Documentation": "https://uddsketch.readthedocs.io",
+    "Issues": "https://github.com/jettify/uddsketch/issues",
+}
 
-setup(name='aioodbc',
-      version=read_version(),
-      description=('ODBC driver for asyncio.'),
-      long_description='\n\n'.join((read('README.rst'), read('CHANGES.txt'))),
-      classifiers=classifiers,
-      platforms=['POSIX'],
-      author="Nikolay Novik",
-      author_email="nickolainovik@gmail.com",
-      url='https://github.com/aio-libs/aioodbc',
-      download_url='https://pypi.python.org/pypi/aioodbc',
-      license='Apache 2',
-      packages=find_packages(),
-      python_requires='>=3.5',
-      install_requires=install_requires,
-      extras_require=extras_require,
-      include_package_data=True)
+
+setup(
+    name="aioodbc",
+    description=("ODBC driver for asyncio."),
+    long_description="\n\n".join((read("README.rst"), read("CHANGES.txt"))),
+    long_description_content_type="text/x-rst",
+    classifiers=classifiers,
+    platforms=["POSIX"],
+    author="Nikolay Novik",
+    author_email="nickolainovik@gmail.com",
+    url="https://github.com/aio-libs/aioodbc",
+    download_url="https://pypi.python.org/pypi/aioodbc",
+    license="Apache 2",
+    packages=find_packages(exclude=("tests",)),
+    python_requires=">=3.7",
+    install_requires=install_requires,
+    setup_requires=[
+        "setuptools>=45",
+        "setuptools_scm",
+        "setuptools_scm_git_archive",
+        "wheel",
+    ],
+    extras_require=extras_require,
+    include_package_data=True,
+    project_urls=project_urls,
+    use_scm_version=True,
+)
