@@ -5,63 +5,50 @@ import warnings
 from functools import partial
 
 import pyodbc
+
 from .cursor import Cursor
 from .utils import _ContextManager, _is_conn_close_error
 
-
-__all__ = ['connect', 'Connection']
-
-
-def connect(*, dsn, autocommit=False, ansi=False, timeout=0, loop=None,
-            executor=None, echo=False, after_created=None, **kwargs):
-    """Accepts an ODBC connection string and returns a new Connection object.
-
-    The connection string can be passed as the string `str`, as a list of
-    keywords,or a combination of the two.  Any keywords except autocommit,
-    ansi, and timeout are simply added to the connection string.
-
-    :param autocommit bool: False or zero, the default, if True or non-zero,
-        the connection is put into ODBC autocommit mode and statements are
-        committed automatically.
-    :param ansi bool: By default, pyodbc first attempts to connect using
-        the Unicode version of SQLDriverConnectW. If the driver returns IM001
-        indicating it does not support the Unicode version, the ANSI version
-        is tried.
-    :param timeout int: An integer login timeout in seconds, used to set
-        the SQL_ATTR_LOGIN_TIMEOUT attribute of the connection. The default is
-         0  which means the database's default timeout, if any, is use
-    :param after_created callable: support customize configuration after
-        connection is connected.  Must be an async unary function, or leave it
-        as None.
-    """
-    return _ContextManager(_connect(dsn=dsn, autocommit=autocommit,
-                           ansi=ansi, timeout=timeout, loop=loop,
-                           executor=executor, echo=echo,
-                           after_created=after_created, **kwargs))
+__all__ = ["connect", "Connection"]
 
 
-async def _connect(*, dsn, autocommit=False, ansi=False, timeout=0, loop=None,
-                   executor=None, echo=False, after_created=None, **kwargs):
-    loop = loop or asyncio.get_event_loop()
-    conn = Connection(dsn=dsn, autocommit=autocommit, ansi=ansi,
-                      timeout=timeout, echo=echo, loop=loop, executor=executor,
-                      after_created=after_created, **kwargs)
-    await conn._connect()
-    return conn
+async def _close_cursor(c: Cursor) -> None:
+    if not c.autocommit:
+        await c.commit()
+    await c.close()
+
+
+async def _close_cursor_on_error(c: Cursor) -> None:
+    await c.rollback()
+    await c.close()
 
 
 class Connection:
-    """ Connection objects manage connections to the database.
+    """Connection objects manage connections to the database.
 
     Connections should only be created by the aioodbc.connect function.
     """
+
     _source_traceback = None
 
-    def __init__(self, *, dsn, autocommit=False, ansi=None,
-                 timeout=0, executor=None, echo=False, loop=None,
-                 after_created=None, **kwargs):
+    def __init__(
+        self,
+        *,
+        dsn,
+        autocommit=False,
+        ansi=None,
+        timeout=0,
+        executor=None,
+        echo=False,
+        loop=None,  # deprecated
+        after_created=None,
+        **kwargs,
+    ):
+        if loop is not None:
+            msg = "Explicit loop is deprecated, and has no effect."
+            warnings.warn(msg, DeprecationWarning, stacklevel=2)
         self._executor = executor
-        self._loop = loop or asyncio.get_event_loop()
+        self._loop = asyncio.get_event_loop()
         self._conn = None
 
         self._timeout = timeout
@@ -72,7 +59,7 @@ class Connection:
         self._echo = echo
         self._posthook = after_created
         self._kwargs = kwargs
-        if loop.get_debug():
+        if self._loop.get_debug():
             self._source_traceback = traceback.extract_stack(sys._getframe(1))
 
     def _execute(self, func, *args, **kwargs):
@@ -81,22 +68,26 @@ class Connection:
         future = self._loop.run_in_executor(self._executor, func)
         return future
 
-    async def _connect(self):
+    async def _connect(self) -> None:
         # create pyodbc connection
-        f = self._execute(pyodbc.connect, self._dsn,
-                          autocommit=self._autocommit, ansi=self._ansi,
-                          timeout=self._timeout,
-                          **self._kwargs)
+        f = self._execute(
+            pyodbc.connect,
+            self._dsn,
+            autocommit=self._autocommit,
+            ansi=self._ansi,
+            timeout=self._timeout,
+            **self._kwargs,
+        )
         self._conn = await f
         if self._posthook is not None:
             await self._posthook(self._conn)
 
     @property
-    def loop(self):
+    def loop(self) -> asyncio.AbstractEventLoop:
         return self._loop
 
     @property
-    def closed(self):
+    def closed(self) -> bool:
         if self._conn:
             return False
         return True
@@ -107,18 +98,25 @@ class Connection:
         connection is in autocommit mode; False otherwise. The default
         is False
         """
+        assert self._conn is not None  # mypy
         return self._conn.autocommit
 
+    @autocommit.setter
+    def autocommit(self, value):
+        assert self._conn is not None  # mypy
+        self._conn.autocommit = value
+
     @property
-    def timeout(self):
+    def timeout(self) -> int:
+        assert self._conn is not None  # mypy
         return self._conn.timeout
 
     @property
-    def last_usage(self):
+    def last_usage(self) -> float:
         return self._last_usage
 
     @property
-    def echo(self):
+    def echo(self) -> bool:
         return self._echo
 
     async def _cursor(self):
@@ -127,10 +125,12 @@ class Connection:
         connection = self
         return Cursor(c, connection, echo=self._echo)
 
-    def cursor(self):
-        return _ContextManager(self._cursor())
+    def cursor(self) -> _ContextManager:
+        return _ContextManager["Cursor"](
+            self._cursor(), _close_cursor, _close_cursor_on_error
+        )
 
-    async def close(self):
+    async def close(self) -> None:
         """Close pyodbc connection"""
         if not self._conn:
             return
@@ -138,19 +138,19 @@ class Connection:
         self._conn = None
         return c
 
-    def commit(self):
+    async def commit(self) -> None:
         """Commit any pending transaction to the database."""
-        fut = self._execute(self._conn.commit)
-        return fut
+        assert self._conn is not None  # mypy
+        await self._execute(self._conn.commit)
 
-    def rollback(self):
+    async def rollback(self) -> None:
         """Causes the database to roll back to the start of any pending
         transaction.
         """
-        fut = self._execute(self._conn.rollback)
-        return fut
+        assert self._conn is not None  # mypy
+        await self._execute(self._conn.rollback)
 
-    async def execute(self, sql, *args):
+    async def execute(self, sql: str, *args) -> Cursor:
         """Create a new Cursor object, call its execute method, and return it.
 
         See Cursor.execute for more details.This is a convenience method
@@ -161,6 +161,7 @@ class Connection:
         :param sql: str, formatted sql statement
         :param args: tuple, arguments for construction of sql statement
         """
+        assert self._conn is not None  # mypy
         try:
             _cursor = await self._execute(self._conn.execute, sql, *args)
             connection = self
@@ -223,13 +224,13 @@ class Connection:
             self._conn.close()
             self._conn = None
 
-            warnings.warn("Unclosed connection {!r}".format(self),
-                          ResourceWarning)
+            warnings.warn(
+                f"Unclosed connection {self!r}", ResourceWarning, stacklevel=1
+            )
 
-            context = {'connection': self,
-                       'message': 'Unclosed connection'}
+            context = {"connection": self, "message": "Unclosed connection"}
             if self._source_traceback is not None:
-                context['source_traceback'] = self._source_traceback
+                context["source_traceback"] = self._source_traceback
             self._loop.call_exception_handler(context)
 
     async def __aenter__(self):
@@ -238,3 +239,91 @@ class Connection:
     async def __aexit__(self, exc_type, exc_val, exc_tb):
         await self.close()
         return
+
+
+async def _disconnect(c: "Connection") -> None:
+    if not c.autocommit:
+        await c.commit()
+    await c.close()
+
+
+async def _disconnect_on_error(c: "Connection") -> None:
+    await c.rollback()
+    await c.close()
+
+
+async def _connect(
+    *,
+    dsn,
+    autocommit=False,
+    ansi=False,
+    timeout=0,
+    executor=None,
+    echo=False,
+    after_created=None,
+    **kwargs,
+):
+    conn = Connection(
+        dsn=dsn,
+        autocommit=autocommit,
+        ansi=ansi,
+        timeout=timeout,
+        echo=echo,
+        loop=None,  # deprecated
+        executor=executor,
+        after_created=after_created,
+        **kwargs,
+    )
+    await conn._connect()
+    return conn
+
+
+def connect(
+    *,
+    dsn,
+    autocommit=False,
+    ansi=False,
+    timeout=0,
+    loop=None,
+    executor=None,
+    echo=False,
+    after_created=None,
+    **kwargs,
+):
+    """Accepts an ODBC connection string and returns a new Connection object.
+
+    The connection string can be passed as the string `str`, as a list of
+    keywords,or a combination of the two.  Any keywords except autocommit,
+    ansi, and timeout are simply added to the connection string.
+
+    :param autocommit bool: False or zero, the default, if True or non-zero,
+        the connection is put into ODBC autocommit mode and statements are
+        committed automatically.
+    :param ansi bool: By default, pyodbc first attempts to connect using
+        the Unicode version of SQLDriverConnectW. If the driver returns IM001
+        indicating it does not support the Unicode version, the ANSI version
+        is tried.
+    :param timeout int: An integer login timeout in seconds, used to set
+        the SQL_ATTR_LOGIN_TIMEOUT attribute of the connection. The default is
+         0  which means the database's default timeout, if any, is use
+    :param after_created callable: support customize configuration after
+        connection is connected.  Must be an async unary function, or leave it
+        as None.
+    """
+    if loop is not None:
+        msg = "Explicit loop is deprecated, and has no effect."
+        warnings.warn(msg, DeprecationWarning, stacklevel=2)
+    return _ContextManager(
+        _connect(
+            dsn=dsn,
+            autocommit=autocommit,
+            ansi=ansi,
+            timeout=timeout,
+            executor=executor,
+            echo=echo,
+            after_created=after_created,
+            **kwargs,
+        ),
+        _disconnect,
+        _disconnect_on_error,
+    )
