@@ -1,13 +1,11 @@
 import asyncio
+
 import aioodbc
 
 
-loop = asyncio.get_event_loop()
-
-
 async def test_pool():
-    dsn = 'Driver=SQLite;Database=sqlite.db'
-    pool = await aioodbc.create_pool(dsn=dsn, loop=loop)
+    dsn = "Driver=SQLite3;Database=sqlite_pool.db"
+    pool = await aioodbc.create_pool(dsn=dsn)
 
     async with pool.acquire() as conn:
         cur = await conn.cursor()
@@ -19,4 +17,5 @@ async def test_pool():
     pool.close()
     await pool.wait_closed()
 
-loop.run_until_complete(test_pool())
+
+asyncio.run(test_pool())
