@@ -1,21 +1,20 @@
-import asyncio
 import gc
 from unittest import mock
 
-import pytest
 import pyodbc
+import pytest
 
 import aioodbc
 
 
-def test_connect(loop, conn):
-    assert conn.loop is loop
+@pytest.mark.asyncio
+async def test_connect(conn):
     assert not conn.autocommit
     assert conn.timeout == 0
     assert not conn.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_connect_hook(connection_maker):
     raw_conn = None
@@ -28,26 +27,17 @@ async def test_connect_hook(connection_maker):
     assert connection._conn == raw_conn
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_basic_cursor(conn):
     cursor = await conn.cursor()
-    sql = 'SELECT 10;'
+    sql = "SELECT 10;"
     await cursor.execute(sql)
     (resp,) = await cursor.fetchone()
     assert resp == 10
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
-@pytest.mark.asyncio
-async def test_default_loop(loop, dsn):
-    asyncio.set_event_loop(loop)
-    conn = await aioodbc.connect(dsn=dsn)
-    assert conn._loop is loop
-    await conn.close()
-
-
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_close_twice(conn):
     await conn.close()
@@ -55,17 +45,17 @@ async def test_close_twice(conn):
     assert conn.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_execute(conn):
-    cur = await conn.execute('SELECT 10;')
+    cur = await conn.execute("SELECT 10;")
     (resp,) = await cur.fetchone()
     await conn.close()
     assert resp == 10
     assert conn.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_getinfo(conn):
     data = await conn.getinfo(pyodbc.SQL_CREATE_TABLE)
@@ -75,15 +65,15 @@ async def test_getinfo(conn):
     assert data in (pg, sqlite, mysql)
 
 
-@pytest.mark.parametrize('db', ['mysql'])
+@pytest.mark.parametrize("db", ["sqlite"])
 @pytest.mark.asyncio
 async def test_output_conversion(conn, table):
     def convert(value):
         # value will be a string.  We'll simply add an X at the
         # beginning at the end.
         if isinstance(value, str):
-            return 'X' + value + 'X'
-        return b'X' + value + b'X'
+            return "X" + value + "X"
+        return b"X" + value + b"X"
 
     await conn.add_output_converter(pyodbc.SQL_VARCHAR, convert)
     cur = await conn.cursor()
@@ -92,25 +82,25 @@ async def test_output_conversion(conn, table):
     await cur.execute("SELECT v FROM t1 WHERE n=3;")
     (value,) = await cur.fetchone()
 
-    assert value in (b'X123.45X', 'X123.45X')
+    assert value in (b"X123.45X", "X123.45X")
 
     # Now clear the conversions and try again. There should be
     # no Xs this time.
     await conn.clear_output_converters()
     await cur.execute("SELECT v FROM t1")
     (value,) = await cur.fetchone()
-    assert value == '123.45'
+    assert value == "123.45"
     await cur.close()
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_autocommit(loop, connection_maker):
+async def test_autocommit(connection_maker):
     conn = await connection_maker(autocommit=True)
     assert conn.autocommit, True
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_rollback(conn):
     assert not conn.autocommit
@@ -123,7 +113,7 @@ async def test_rollback(conn):
     await cur.execute("INSERT INTO t1 VALUES (1, '123.45');")
     await cur.execute("SELECT v FROM t1")
     (value,) = await cur.fetchone()
-    assert value == '123.45'
+    assert value == "123.45"
 
     await conn.rollback()
     await cur.execute("SELECT v FROM t1;")
@@ -135,12 +125,12 @@ async def test_rollback(conn):
     await conn.close()
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_custom_executor(loop, dsn, executor):
-    conn = await aioodbc.connect(dsn=dsn, executor=executor, loop=loop)
+async def test_custom_executor(dsn, executor):
+    conn = await aioodbc.connect(dsn=dsn, executor=executor)
     assert conn._executor is executor
-    cur = await conn.execute('SELECT 10;')
+    cur = await conn.execute("SELECT 10;")
     (resp,) = await cur.fetchone()
     await conn.close()
     assert resp == 10
@@ -148,14 +138,14 @@ async def test_custom_executor(loop, dsn, executor):
 
 
 @pytest.mark.asyncio
-async def test_dataSources(loop, executor):
-    data = await aioodbc.dataSources(loop, executor)
+async def test_dataSources(executor):
+    data = await aioodbc.dataSources(executor)
     assert isinstance(data, dict)
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_connection_simple_with(loop, conn):
+async def test_connection_simple_with(conn):
     assert not conn.closed
     async with conn:
         pass
@@ -163,14 +153,14 @@ async def test_connection_simple_with(loop, conn):
     assert conn.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_connect_context_manager(loop, dsn):
-    async with aioodbc.connect(dsn=dsn, loop=loop, echo=True) as conn:
+async def test_connect_context_manager(dsn):
+    async with aioodbc.connect(dsn=dsn, echo=True) as conn:
         assert not conn.closed
         assert conn.echo
 
-        cur = await conn.execute('SELECT 10;')
+        cur = await conn.execute("SELECT 10;")
         assert cur.echo
         (resp,) = await cur.fetchone()
         assert resp == 10
@@ -179,11 +169,10 @@ async def test_connect_context_manager(loop, dsn):
     assert conn.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test___del__(loop, dsn, recwarn, executor):
-    return
-    conn = await aioodbc.connect(dsn=dsn, loop=loop, executor=executor)
+    conn = await aioodbc.connect(dsn=dsn, executor=executor)
     exc_handler = mock.Mock()
     loop.set_exception_handler(exc_handler)
 
@@ -192,8 +181,8 @@ async def test___del__(loop, dsn, recwarn, executor):
     w = recwarn.pop()
     assert issubclass(w.category, ResourceWarning)
 
-    msg = {'connection': mock.ANY,  # conn was deleted
-           'message': 'Unclosed connection'}
-    if loop.get_debug():
-        msg['source_traceback'] = mock.ANY
+    msg = {
+        "connection": mock.ANY,  # conn was deleted
+        "message": "Unclosed connection",
+    }
     exc_handler.assert_called_with(loop, msg)
