@@ -2,14 +2,15 @@ import asyncio
 import time
 
 import pytest
-import aioodbc
-from aioodbc import Pool, Connection
 from pyodbc import Error
+
+import aioodbc
+from aioodbc import Connection, Pool
 
 
 @pytest.mark.asyncio
-async def test_create_pool(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn)
+async def test_create_pool(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn)
     assert isinstance(pool, Pool)
     assert 10 == pool.minsize
     assert 10 == pool.maxsize
@@ -19,8 +20,8 @@ async def test_create_pool(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_create_pool2(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, maxsize=20)
+async def test_create_pool2(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, maxsize=20)
     assert isinstance(pool, Pool)
     assert 10 == pool.minsize
     assert 20 == pool.maxsize
@@ -28,7 +29,7 @@ async def test_create_pool2(loop, pool_maker, dsn):
     assert 10 == pool.freesize
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
 async def test_acquire(pool):
     conn = await pool.acquire()
@@ -36,7 +37,7 @@ async def test_acquire(pool):
         assert isinstance(conn, Connection)
         assert not conn.closed
         cur = await conn.cursor()
-        await cur.execute('SELECT 1')
+        await cur.execute("SELECT 1")
         val = await cur.fetchone()
         assert (1,) == tuple(val)
     finally:
@@ -55,25 +56,28 @@ async def test_release(pool):
     assert not pool._used
 
 
+@pytest.mark.skip(reason="PG fixture needs update.")
 @pytest.mark.asyncio
-async def test_op_error_release(loop, pool_maker, pg_server_local):
-    pool = await pool_maker(loop, dsn=pg_server_local['dsn'], autocommit=True)
+async def test_op_error_release(pool_maker, pg_server_local):
+    pool = await pool_maker(dsn=pg_server_local["dsn"], autocommit=True)
 
     async with pool.acquire() as conn:
+
         async def execute():
             start = time.time()
 
             while time.time() - start < 20:
-                await conn.execute('SELECT 1; SELECT pg_sleep(1);')
+                await conn.execute("SELECT 1; SELECT pg_sleep(1);")
 
         async def _kill_conn():
             await asyncio.sleep(2)
-            await pg_server_local['container'].kill()
-            await pg_server_local['container'].delete(v=True, force=True)
-            pg_server_local['container'] = None
+            await pg_server_local["container"].kill()
+            await pg_server_local["container"].delete(v=True, force=True)
+            pg_server_local["container"] = None
 
         result = await asyncio.gather(
-            _kill_conn(), execute(), return_exceptions=True)
+            _kill_conn(), execute(), return_exceptions=True
+        )
         exc = result[1]
         assert isinstance(exc, Error)
 
@@ -116,8 +120,8 @@ async def test_clear(pool):
 
 
 @pytest.mark.asyncio
-async def test_initial_empty(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=0)
+async def test_initial_empty(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=0)
 
     assert 10 == pool.maxsize
     assert 0 == pool.minsize
@@ -151,8 +155,8 @@ async def test_initial_empty(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_parallel_tasks(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=0, maxsize=2)
+async def test_parallel_tasks(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=0, maxsize=2)
 
     assert 2 == pool.maxsize
     assert 0 == pool.minsize
@@ -162,7 +166,7 @@ async def test_parallel_tasks(loop, pool_maker, dsn):
     fut1 = pool.acquire()
     fut2 = pool.acquire()
 
-    conn1, conn2 = await asyncio.gather(fut1, fut2, loop=loop)
+    conn1, conn2 = await asyncio.gather(fut1, fut2)
     assert 2 == pool.size
     assert 0 == pool.freesize
     assert {conn1, conn2} == pool._used
@@ -184,15 +188,14 @@ async def test_parallel_tasks(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_parallel_tasks_more(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=0, maxsize=3)
+async def test_parallel_tasks_more(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=0, maxsize=3)
 
     fut1 = pool.acquire()
     fut2 = pool.acquire()
     fut3 = pool.acquire()
 
-    conn1, conn2, conn3 = await asyncio.gather(fut1, fut2, fut3,
-                                               loop=loop)
+    conn1, conn2, conn3 = await asyncio.gather(fut1, fut2, fut3)
     assert 3 == pool.size
     assert 0 == pool.freesize
     assert {conn1, conn2, conn3} == pool._used
@@ -223,24 +226,15 @@ async def test_parallel_tasks_more(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_default_loop(loop, dsn):
-    pool = await aioodbc.create_pool(dsn=dsn)
-    assert pool._loop is loop
-    pool.close()
-    await pool.wait_closed()
-
-
-@pytest.mark.asyncio
-async def test__fill_free(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=1)
+async def test__fill_free(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=1)
 
     first_conn = await pool.acquire()
     try:
         assert 0 == pool.freesize
         assert 1 == pool.size
 
-        conn = await asyncio.wait_for(pool.acquire(), timeout=0.5,
-                                      loop=loop)
+        conn = await asyncio.wait_for(pool.acquire(), timeout=0.5)
         assert 0 == pool.freesize
         assert 2 == pool.size
         await pool.release(conn)
@@ -253,8 +247,8 @@ async def test__fill_free(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_connect_from_acquire(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=0)
+async def test_connect_from_acquire(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=0)
 
     assert 0 == pool.freesize
     assert 0 == pool.size
@@ -269,16 +263,12 @@ async def test_connect_from_acquire(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_pool_with_connection_recycling(loop, pool_maker, dsn):
-    pool = await pool_maker(loop,
-                            dsn=dsn,
-                            minsize=1,
-                            maxsize=1,
-                            pool_recycle=3)
+async def test_pool_with_connection_recycling(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=1, maxsize=1, pool_recycle=3)
     async with pool.acquire() as conn:
         conn1 = conn
 
-    await asyncio.sleep(5, loop=loop)
+    await asyncio.sleep(5)
 
     assert 1 == pool.freesize
     async with pool.acquire() as conn:
@@ -288,8 +278,8 @@ async def test_pool_with_connection_recycling(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_concurrency(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=2, maxsize=4)
+async def test_concurrency(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=2, maxsize=4)
 
     c1 = await pool.acquire()
     c2 = await pool.acquire()
@@ -300,18 +290,17 @@ async def test_concurrency(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_invalid_minsize_and_maxsize(loop, dsn):
+async def test_invalid_minsize_and_maxsize(dsn):
     with pytest.raises(ValueError):
-        await aioodbc.create_pool(dsn=dsn, loop=loop, minsize=-1)
+        await aioodbc.create_pool(dsn=dsn, minsize=-1)
 
     with pytest.raises(ValueError):
-        await aioodbc.create_pool(dsn=dsn, loop=loop, minsize=5,
-                                  maxsize=2)
+        await aioodbc.create_pool(dsn=dsn, minsize=5, maxsize=2)
 
 
 @pytest.mark.asyncio
-async def test_true_parallel_tasks(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, minsize=0, maxsize=1)
+async def test_true_parallel_tasks(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, minsize=0, maxsize=1)
 
     assert 1 == pool.maxsize
     assert 0 == pool.minsize
@@ -328,20 +317,20 @@ async def test_true_parallel_tasks(loop, pool_maker, dsn):
         conn = await pool.acquire()
         maxsize = max(maxsize, pool.size)
         minfreesize = min(minfreesize, pool.freesize)
-        await asyncio.sleep(0.01, loop=loop)
+        await asyncio.sleep(0.01)
         await pool.release(conn)
         maxsize = max(maxsize, pool.size)
         minfreesize = min(minfreesize, pool.freesize)
 
-    await asyncio.gather(inner(), inner(), loop=loop)
+    await asyncio.gather(inner(), inner())
 
     assert 1 == maxsize
     assert 0 == minfreesize
 
 
 @pytest.mark.asyncio
-async def test_cannot_acquire_after_closing(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn)
+async def test_cannot_acquire_after_closing(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn)
 
     pool.close()
 
@@ -350,8 +339,8 @@ async def test_cannot_acquire_after_closing(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_wait_closed(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn)
+async def test_wait_closed(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn)
 
     c1 = await pool.acquire()
     c2 = await pool.acquire()
@@ -361,26 +350,23 @@ async def test_wait_closed(loop, pool_maker, dsn):
     ops = []
 
     async def do_release(conn):
-        await asyncio.sleep(0, loop=loop)
+        await asyncio.sleep(0)
         await pool.release(conn)
-        ops.append('release')
+        ops.append("release")
 
     async def wait_closed():
         await pool.wait_closed()
-        ops.append('wait_closed')
+        ops.append("wait_closed")
 
     pool.close()
-    await asyncio.gather(wait_closed(),
-                         do_release(c1),
-                         do_release(c2),
-                         loop=loop)
-    assert ['release', 'release', 'wait_closed'] == ops
+    await asyncio.gather(wait_closed(), do_release(c1), do_release(c2))
+    assert ["release", "release", "wait_closed"] == ops
     assert 0 == pool.freesize
 
 
 @pytest.mark.asyncio
-async def test_echo(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn, echo=True)
+async def test_echo(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn, echo=True)
 
     assert pool.echo
     conn = await pool.acquire()
@@ -389,8 +375,8 @@ async def test_echo(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_release_closed_connection(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn)
+async def test_release_closed_connection(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn)
 
     conn = await pool.acquire()
     await conn.close()
@@ -400,8 +386,8 @@ async def test_release_closed_connection(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_wait_closing_on_not_closed(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn)
+async def test_wait_closing_on_not_closed(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn)
 
     with pytest.raises(RuntimeError):
         await pool.wait_closed()
@@ -409,23 +395,22 @@ async def test_wait_closing_on_not_closed(loop, pool_maker, dsn):
 
 
 @pytest.mark.asyncio
-async def test_close_with_acquired_connections(loop, pool_maker, dsn):
-    pool = await pool_maker(loop, dsn=dsn)
+async def test_close_with_acquired_connections(pool_maker, dsn):
+    pool = await pool_maker(dsn=dsn)
 
     conn = await pool.acquire()
     pool.close()
 
     with pytest.raises(asyncio.TimeoutError):
-        await asyncio.wait_for(pool.wait_closed(), 0.1, loop=loop)
+        await asyncio.wait_for(pool.wait_closed(), 0.1)
     await conn.close()
     await pool.release(conn)
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_pool_with_executor(loop, pool_maker, dsn, executor):
-    pool = await pool_maker(
-        loop, executor=executor, dsn=dsn, minsize=2, maxsize=2)
+async def test_pool_with_executor(pool_maker, dsn, executor):
+    pool = await pool_maker(executor=executor, dsn=dsn, minsize=2, maxsize=2)
 
     conn = await pool.acquire()
     try:
@@ -433,7 +418,7 @@ async def test_pool_with_executor(loop, pool_maker, dsn, executor):
         assert not conn.closed
         assert conn._executor is executor
         cur = await conn.cursor()
-        await cur.execute('SELECT 1')
+        await cur.execute("SELECT 1")
         val = await cur.fetchone()
         assert (1,) == tuple(val)
     finally:
@@ -444,30 +429,30 @@ async def test_pool_with_executor(loop, pool_maker, dsn, executor):
     await pool.wait_closed()
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_pool_context_manager(loop, pool):
+async def test_pool_context_manager(pool):
     assert not pool.closed
     async with pool:
         assert not pool.closed
     assert pool.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_pool_context_manager2(loop, pool):
+async def test_pool_context_manager2(pool):
     async with pool.acquire() as conn:
         assert not conn.closed
         cur = await conn.cursor()
-        await cur.execute('SELECT 1')
+        await cur.execute("SELECT 1")
         val = await cur.fetchone()
         assert (1,) == tuple(val)
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_all_context_managers(dsn, loop, executor):
-    kw = {'dsn': dsn, 'loop': loop, 'executor': executor}
+async def test_all_context_managers(dsn, executor):
+    kw = {"dsn": dsn, "executor": executor}
     async with aioodbc.create_pool(**kw) as pool:
         async with pool.acquire() as conn:
             async with conn.cursor() as cur:
@@ -475,7 +460,7 @@ async def test_all_context_managers(dsn, loop, executor):
                 assert not conn.closed
                 assert not cur.closed
 
-                await cur.execute('SELECT 1')
+                await cur.execute("SELECT 1")
                 val = await cur.fetchone()
                 assert (1,) == tuple(val)
 
@@ -484,18 +469,18 @@ async def test_all_context_managers(dsn, loop, executor):
     assert cur.closed
 
 
-@pytest.mark.parametrize('db', pytest.db_list)
+@pytest.mark.parametrize("db", pytest.db_list)
 @pytest.mark.asyncio
-async def test_context_manager_aexit(loop, connection_maker):
+async def test_context_manager_aexit(connection_maker):
     async def aexit_conntex_managet(conn):
         # commit on exit if no error
-        params = (1, '123.45')
+        params = (1, "123.45")
         async with conn.cursor() as cur:
             await cur.execute("CREATE TABLE cmt1(n int, v VARCHAR(10))")
             await cur.execute("INSERT INTO cmt1 VALUES (?,?);", params)
         async with conn.cursor() as cur:
             await cur.execute("SELECT v FROM cmt1 WHERE n=1;")
-            (value, ) = await cur.fetchone()
+            (value,) = await cur.fetchone()
             assert value == params[1]
 
         # rollback on exit if error
